+++ packages/comment2/src/node/plugin.ts
@@ -6,6 +6,7 @@ import {
   addViteOptimizeDepsExclude,
   addViteOptimizeDepsInclude,
   addViteSsrExternal,
+  addViteSsrNoExternal,
   checkVersion,
   getLocales,
 } from "vuepress-shared/node";
@@ -25,7 +26,7 @@ export const commentPlugin =
     // TODO: Remove this in v2 stable
     if (legacy)
       convertOptions(options as CommentPluginOptions & Record<string, unknown>);
-    checkVersion(app, PLUGIN_NAME, "2.0.0-beta.66");
+    checkVersion(app, PLUGIN_NAME, "2.0.0-beta.67");
 
     if (app.env.isDebug) logger.info("Options:", options);
 
@@ -93,6 +94,8 @@ export const commentPlugin =
             break;
           }
         }
+
+        addViteSsrNoExternal(bundlerOptions, app, "vuepress-sahre");
       },
 
       clientConfigFile: path.resolve(__dirname, "../client/config.js"),