+++ packages/copyright2/src/node/compact/convert.ts
@@ -1,0 +1,13 @@
+import { deprecatedLogger } from "./utils.js";
+import type { CopyrightOptions } from "../options.js";
+
+/** @deprecated */
+export const convertOptions = (
+  options: CopyrightOptions & Record<string, unknown>,
+): void => {
+  deprecatedLogger({
+    options,
+    deprecatedOption: "triggerWords",
+    newOption: "triggerLength",
+  });
+};