+++ packages/copyright2/src/node/plugins.ts
@@ -1,7 +1,8 @@
 import type { Page, PluginFunction } from "@vuepress/core";
-import { getDirname, path } from "@vuepress/utils";
+import { colors, getDirname, path } from "@vuepress/utils";
 import { checkVersion, getLocales, isFunction } from "vuepress-shared/node";
 
+import { convertOptions } from "./compact/index.js";
 import { copyrightLocales } from "./locales.js";
 import type { CopyrightOptions } from "./options.js";
 import { PLUGIN_NAME, logger } from "./utils.js";
@@ -10,9 +11,13 @@ import type { CopyrightPluginPageData } from "../shared/index.js";
 const __dirname = getDirname(import.meta.url);
 
 export const copyrightPlugin =
-  (options: CopyrightOptions): PluginFunction =>
+  (options: CopyrightOptions, legacy = true): PluginFunction =>
   (app) => {
-    checkVersion(app, PLUGIN_NAME, "2.0.0-beta.66");
+    // TODO: Remove this in v2 stable
+    if (legacy)
+      convertOptions(options as CopyrightOptions & Record<string, unknown>);
+
+    checkVersion(app, PLUGIN_NAME, "2.0.0-beta.67");
 
     if (app.env.isDebug) logger.info("Options:", options);
 
@@ -23,7 +28,8 @@ export const copyrightPlugin =
       disableCopy = false,
       disableSelection = false,
       global = false,
-      triggerWords = 100,
+      triggerLength = 100,
+      maxLength = 0,
     } = options;
 
     const locales = getLocales({
@@ -42,13 +48,30 @@ export const copyrightPlugin =
         COPYRIGHT_DISABLE_COPY: disableCopy,
         COPYRIGHT_DISABLE_SELECTION: disableSelection,
         COPYRIGHT_LOCALES: locales,
-        COPYRIGHT_TRIGGER_WORDS: triggerWords,
+        COPYRIGHT_MAX_LENGTH: maxLength,
+        COPYRIGHT_TRIGGER_LENGTH: triggerLength,
       }),
 
       extendsPage: (page: Page<Partial<CopyrightPluginPageData>>): void => {
         const authorText = isFunction(author) ? author(page) : author;
 
         const licenseText = isFunction(license) ? license(page) : license;
+
+        if (page.frontmatter["triggerWords"]) {
+          logger.warn(
+            `The ${colors.cyan(
+              "triggerWords",
+            )} option is deprecated, use ${colors.cyan(
+              "triggerLength",
+            )} instead${
+              page.filePathRelative
+                ? `, found in ${page.filePathRelative}`
+                : "."
+            }`,
+          );
+
+          page.frontmatter["triggerLength"] = page.frontmatter["triggerWords"];
+        }
 
         page.data.copyright = {
           ...(authorText ? { author: authorText } : {}),